feat: replace old json-based lambdas to db-based lambdas
