+++ product-service/common/constants.ts
