+++ product-service/create-db.sql
@@ -1,26 +1,27 @@
---create extension if not exists "uuid-ossp";
-
 create table products (
 	id uuid primary key default uuid_generate_v4(),
 	title text not null,
 	description text,
-	price integer
+	author text,
+	image text,
+	price real
 );
 
 create table stocks (
-	product_id uuid references products(id),
-	count integer
+	product_id uuid,
+	count integer,
+	foreign key (product_id) references products(id) on delete cascade
 );
 
-insert into products (id, title, description, price) values
-	('25f87050-dea3-4db7-9381-9b66a0cd51c4', 'Рефакторинг. Улучшение проекта существующего кода', 'Классическая книга про практикам программирования от Мартина Фаулера', 2180.4),
-	('9137d163-a173-4bc2-9dc7-d2df27ea3b7b', 'Чистый код. Создание анализ и рефакторинг', 'Тоже классика разработки от дядюшки Боба Мартина', 504),
-	('75d4707a-1c37-49e9-b531-e3967f66ba10', 'Высоконагруженные приложения. Программирование масштабирование поддержка', 'Книга по разработке высоконагруженных систем', 1576),
-	('6a983db2-96cb-4581-ab39-2e1b0a233d49', 'Kubernetes в действии', 'Система контейнерной оркестровки Kubernetes', 1871),
-	('0cc51f32-9e97-4fd5-866a-7bdb2d011ba3', 'Стратегии решения математических задач. Различные подходы к типовым задачам', 'В этой книге рассказывается о десяти различных стратегиях решения задач', 344),	
-	('5041bb4e-44c2-46a2-9509-b83e23a5fd70', 'Выразительный JavaScript. Современное веб-программирование', 'Получайте опыт и изучайте язык на множестве примеров, выполняя упражнения и учебные проекты', 1169),
-	('285bd041-8c39-478c-a40b-5e53685937bb', 'Код. Тайный язык информатики', 'Эта книга — азбука компьютерных технологий', 184.9),
-	('a499223f-3b7a-4249-abbb-10b1c8a6848f', 'Дизайн и эволюция С++', 'В книге, написанной создателем языка С++ Бьерпом Страуструпом, представлено описание процесса проектирования и разработки языка программирования С++', 815);
+insert into products (id, title, description, author, image, price) values
+	('25f87050-dea3-4db7-9381-9b66a0cd51c4', 'Р РµС„Р°РєС‚РѕСЂРёРЅРі. РЈР»СѓС‡С€РµРЅРёРµ РїСЂРѕРµРєС‚Р° СЃСѓС‰РµСЃС‚РІСѓСЋС‰РµРіРѕ РєРѕРґР°', 'РљР»Р°СЃСЃРёС‡РµСЃРєР°СЏ РєРЅРёРіР° РїСЂРѕ РїСЂР°РєС‚РёРєР°Рј РїСЂРѕРіСЂР°РјРјРёСЂРѕРІР°РЅРёСЏ РѕС‚ РњР°СЂС‚РёРЅР° Р¤Р°СѓР»РµСЂР°', 'Р¤Р°СѓР»РµСЂ РњР°СЂС‚РёРЅ', 'https://cdn1.ozone.ru/multimedia/1019869511.jpg', 2180.4),
+	('9137d163-a173-4bc2-9dc7-d2df27ea3b7b', 'Р§РёСЃС‚С‹Р№ РєРѕРґ. РЎРѕР·РґР°РЅРёРµ Р°РЅР°Р»РёР· Рё СЂРµС„Р°РєС‚РѕСЂРёРЅРі', 'РўРѕР¶Рµ РєР»Р°СЃСЃРёРєР° СЂР°Р·СЂР°Р±РѕС‚РєРё РѕС‚ РґСЏРґСЋС€РєРё Р‘РѕР±Р° РњР°СЂС‚РёРЅР°', 'РњР°СЂС‚РёРЅ Р РѕР±РµСЂС‚', 'https://cdn1.ozone.ru/multimedia/1026061260.jpg', 504),
+	('75d4707a-1c37-49e9-b531-e3967f66ba10', 'Р’С‹СЃРѕРєРѕРЅР°РіСЂСѓР¶РµРЅРЅС‹Рµ РїСЂРёР»РѕР¶РµРЅРёСЏ. РџСЂРѕРіСЂР°РјРјРёСЂРѕРІР°РЅРёРµ РјР°СЃС€С‚Р°Р±РёСЂРѕРІР°РЅРёРµ РїРѕРґРґРµСЂР¶РєР°', 'РљРЅРёРіР° РїРѕ СЂР°Р·СЂР°Р±РѕС‚РєРµ РІС‹СЃРѕРєРѕРЅР°РіСЂСѓР¶РµРЅРЅС‹С… СЃРёСЃС‚РµРј', 'РљР»РµРїРїРјР°РЅ РњР°СЂС‚РёРЅ', 'https://cdn1.ozone.ru/s3/multimedia-r/6007449903.jpg', 1576),
+	('6a983db2-96cb-4581-ab39-2e1b0a233d49', 'Kubernetes РІ РґРµР№СЃС‚РІРёРё', 'РЎРёСЃС‚РµРјР° РєРѕРЅС‚РµР№РЅРµСЂРЅРѕР№ РѕСЂРєРµСЃС‚СЂРѕРІРєРё Kubernetes', 'Р›СѓРєС€Р° РњР°СЂРєРѕ', 'https://cdn1.ozone.ru/multimedia/1026091236.jpg', 1871),
+	('0cc51f32-9e97-4fd5-866a-7bdb2d011ba3', 'РЎС‚СЂР°С‚РµРіРёРё СЂРµС€РµРЅРёСЏ РјР°С‚РµРјР°С‚РёС‡РµСЃРєРёС… Р·Р°РґР°С‡. Р Р°Р·Р»РёС‡РЅС‹Рµ РїРѕРґС…РѕРґС‹ Рє С‚РёРїРѕРІС‹Рј Р·Р°РґР°С‡Р°Рј', 'Р’ СЌС‚РѕР№ РєРЅРёРіРµ СЂР°СЃСЃРєР°Р·С‹РІР°РµС‚СЃСЏ Рѕ РґРµСЃСЏС‚Рё СЂР°Р·Р»РёС‡РЅС‹С… СЃС‚СЂР°С‚РµРіРёСЏС… СЂРµС€РµРЅРёСЏ Р·Р°РґР°С‡', 'РљСЂСѓС‚РёРє РЎС‚РёРІРµРЅ', 'https://cdn1.ozone.ru/multimedia/1022149198.jpg', 344),	
+	('5041bb4e-44c2-46a2-9509-b83e23a5fd70', 'Р’С‹СЂР°Р·РёС‚РµР»СЊРЅС‹Р№ JavaScript. РЎРѕРІСЂРµРјРµРЅРЅРѕРµ РІРµР±-РїСЂРѕРіСЂР°РјРјРёСЂРѕРІР°РЅРёРµ', 'РџРѕР»СѓС‡Р°Р№С‚Рµ РѕРїС‹С‚ Рё РёР·СѓС‡Р°Р№С‚Рµ СЏР·С‹Рє РЅР° РјРЅРѕР¶РµСЃС‚РІРµ РїСЂРёРјРµСЂРѕРІ, РІС‹РїРѕР»РЅСЏСЏ СѓРїСЂР°Р¶РЅРµРЅРёСЏ Рё СѓС‡РµР±РЅС‹Рµ РїСЂРѕРµРєС‚С‹', 'РҐР°РІРµСЂР±РµРєРµ РњР°СЂРµР№РЅ', 'https://cdn1.ozone.ru/s3/multimedia-t/6000018281.jpg', 1169),
+	('285bd041-8c39-478c-a40b-5e53685937bb', 'РљРѕРґ. РўР°Р№РЅС‹Р№ СЏР·С‹Рє РёРЅС„РѕСЂРјР°С‚РёРєРё', 'Р­С‚Р° РєРЅРёРіР° вЂ” Р°Р·Р±СѓРєР° РєРѕРјРїСЊСЋС‚РµСЂРЅС‹С… С‚РµС…РЅРѕР»РѕРіРёР№', 'РџРµС‚С†РѕР»СЊРґ Р§Р°СЂР»СЊР·', 'https://cdn1.ozone.ru/multimedia/1004493926.jpg', 184.9),
+	('a499223f-3b7a-4249-abbb-10b1c8a6848f', 'Р”РёР·Р°Р№РЅ Рё СЌРІРѕР»СЋС†РёСЏ РЎ++', 'Р’ РєРЅРёРіРµ, РЅР°РїРёСЃР°РЅРЅРѕР№ СЃРѕР·РґР°С‚РµР»РµРј СЏР·С‹РєР° РЎ++ Р‘СЊРµСЂРїРѕРј РЎС‚СЂР°СѓСЃС‚СЂСѓРїРѕРј, РїСЂРµРґСЃС‚Р°РІР»РµРЅРѕ РѕРїРёСЃР°РЅРёРµ РїСЂРѕС†РµСЃСЃР° РїСЂРѕРµРєС‚РёСЂРѕРІР°РЅРёСЏ Рё СЂР°Р·СЂР°Р±РѕС‚РєРё СЏР·С‹РєР° РїСЂРѕРіСЂР°РјРјРёСЂРѕРІР°РЅРёСЏ РЎ++', 'РЎС‚СЂР°СѓСЃС‚СЂСѓРї Р‘СЊРµСЂРЅ', 'https://cdn1.ozone.ru/multimedia/1014964320.jpg', 815);
 
 insert into stocks (product_id, count) values
 	('25f87050-dea3-4db7-9381-9b66a0cd51c4', 5),
@@ -30,7 +31,5 @@ insert into stocks (product_id, count) values
 	('0cc51f32-9e97-4fd5-866a-7bdb2d011ba3', 10),	
 	('5041bb4e-44c2-46a2-9509-b83e23a5fd70', 15),
 	('285bd041-8c39-478c-a40b-5e53685937bb', 0),
-	('a499223f-3b7a-4249-abbb-10b1c8a6848f', 3);	
-
---drop table products;
---drop table stocks;
+	('a499223f-3b7a-4249-abbb-10b1c8a6848f', 3);
+
