+++ product-service/entities/product.ts
