+++ product-service/handlers/getProducts.ts
@@ -2,7 +2,7 @@ import { APIGatewayProxyHandler, APIGatewayProxyResult } from 'aws-lambda';
 import 'source-map-support/register';
 
 import { Client } from 'pg';
-import { corsHeaders } from './constants';
+import { corsHeaders } from '../common/constants';
 
 const { PG_HOST, PG_PORT, PG_DATABASE, PG_USERNAME, PG_PASSWORD } = process.env;
 
@@ -18,30 +18,23 @@ const dbOptions = {
   connectionTimeoutMillis: 5000,
 };
 
-export const pgGetProductById: APIGatewayProxyHandler = async (event) => {
-  const id = event.pathParameters.id;
-  if (!id) return errorIdIsNotDefined();
-
+export const getProducts: APIGatewayProxyHandler = async () => {
   const client = new Client(dbOptions);
   await client.connect();
 
   try {
     const { rows: products } = await client.query(`
         select p.id, p.title, p.description, p.price, s.count 
-        from products as p inner join stocks as s on p.id = s.product_id
-        where p.id = $1;   
-    `, [id]);
+        from products as p inner join stocks as s on p.id = s.product_id;   
+    `);
 
-    if (products.length == 0) return errorProductNotFound();
-
-    const product = products[0];
-    console.log(product);
+    console.log(products);
 
     return {
       statusCode: 200,
       headers: corsHeaders,
       body: JSON.stringify({
-        product,
+        products,
       }),
     };
   } catch (e) {
@@ -51,27 +44,7 @@ export const pgGetProductById: APIGatewayProxyHandler = async (event) => {
   }
 };
 
-function errorIdIsNotDefined(): Promise<APIGatewayProxyResult> {
-  return Promise.resolve({
-    statusCode: 400,
-    body: JSON.stringify({
-      message: 'product Id is not defined in request',
-    }),
-  });
-}
-
-function errorProductNotFound(): Promise<APIGatewayProxyResult> {
-  return Promise.resolve({
-    statusCode: 404,
-    body: JSON.stringify({
-      message: 'product is not found',
-    }),
-  });
-}
-
-function errorFetchProducts(e: {
-  message: any;
-}): Promise<APIGatewayProxyResult> {
+function errorFetchProducts(e: { message: any; }): Promise<APIGatewayProxyResult> {
   return Promise.resolve({
     statusCode: 500,
     body: JSON.stringify({
